// Backend/src/db/types/public/Estados.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { PaisesCodPais } from './Paises';

/** Identifier type for public.Estados */
export type EstadosCodEstado = number & { __brand: 'public.Estados' };

/** Represents the table public.Estados */
export default interface Estados {
  codEstado: EstadosCodEstado;

  codPais: PaisesCodPais;

  uf: string;

  estado: string;

  dtCriacao: Date;

  dtEdicao: Date;

  codUser: number | null;

  isActive: boolean;
}

/** Represents the initializer for the table public.Estados */
export interface EstadosInitializer {
  codEstado?: EstadosCodEstado;

  codPais: PaisesCodPais;

  uf: string;

  estado: string;

  /** Default value: CURRENT_TIMESTAMP */
  dtCriacao?: Date;

  /** Default value: CURRENT_TIMESTAMP */
  dtEdicao?: Date;

  codUser?: number | null;

  /** Default value: true */
  isActive?: boolean;
}

/** Represents the mutator for the table public.Estados */
export interface EstadosMutator {
  codEstado?: EstadosCodEstado;

  codPais?: PaisesCodPais;

  uf?: string;

  estado?: string;

  dtCriacao?: Date;

  dtEdicao?: Date;

  codUser?: number | null;

  isActive?: boolean;
}